+++ reader/shared/src/test/scala/chester/reader/ParserTest.scala
@@ -15,8 +15,8 @@ class ParserTest extends FunSuite {
         meta.sourcePos match {
           case Some(pos) =>
             assertEquals(pos.fileName, "testFile")
-            assertEquals(pos.range.start.line, 0)
-            assertEquals(pos.range.start.column.i, 0)
+            assertEquals(pos.range.start.line.toInt, 0)
+            assertEquals(pos.range.start.column.i.toInt, 0)
           case None => fail("Source position not found")
         }
       case _ => fail("Parsing failed")
@@ -32,8 +32,8 @@ class ParserTest extends FunSuite {
         meta.sourcePos match {
           case Some(pos) =>
             assertEquals(pos.fileName, "testFile")
-            assertEquals(pos.range.start.line, 0)
-            assertEquals(pos.range.start.column.i, 0)
+            assertEquals(pos.range.start.line.toInt, 0)
+            assertEquals(pos.range.start.column.i.toInt, 0)
           case None => fail("Source position not found")
         }
       case _ => fail("Parsing failed")
